+++ railties/lib/rails/commands/routes/routes_command.rb
@@ -8,8 +8,14 @@ module Rails
       class_option :controller, aliases: "-c", desc: "Filter by a specific controller, e.g. PostsController or Admin::PostsController."
       class_option :grep, aliases: "-g", desc: "Grep routes by a specific pattern."
       class_option :expanded, type: :boolean, aliases: "-E", desc: "Print routes expanded vertically with parts explained."
+      class_option :separated, type: :boolean, aliases: "-s", desc: "Print routes separated by empty line between each controller routes."
 
       def perform(*)
+        if options[:expanded] && options[:separated]
+          say "WARNING: rails routes options cannot be both expanded (-E) and separated (-s).", :red
+          exit 1
+        end
+
         require_application_and_environment!
         require "action_dispatch/routing/inspector"
 
@@ -22,8 +28,10 @@ module Rails
         end
 
         def formatter
-          if options.key?("expanded")
+          if options[:expanded]
             ActionDispatch::Routing::ConsoleFormatter::Expanded.new
+          elsif options[:separated]
+            ActionDispatch::Routing::ConsoleFormatter::Separated.new
           else
             ActionDispatch::Routing::ConsoleFormatter::Sheet.new
           end